✨ : DRDC Service
Former-commit-id: 92ca5e6954de845af26320182a6b1318382c446d

+++ backend/nhdream/src/main/java/com/ssafy/nhdream/domain/voucher/dto/VoucherTokenResDto.java
@@ -3,10 +3,12 @@ package com.ssafy.nhdream.domain.voucher.dto;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 
+import java.math.BigDecimal;
+
 @Data
 @Slf4j
 public class VoucherTokenResDto {
 
-    private int amount;
+    private BigDecimal amount;
 
 }
